refactor : move the generate code to generate.js
The separation between the generation of the CompressedObjects and the
actual use of them will help for the next commit, adding generateAsync().
