+++ lib/generate.js
@@ -1,0 +1,308 @@
+'use strict';
+
+var utils = require('./utils');
+var utf8 = require('./utf8');
+var crc32 = require('./crc32');
+var signature = require('./signature');
+var base64 = require('./base64');
+var compressions = require('./compressions');
+var StringWriter = require('./writer/stringWriter');
+var Uint8ArrayWriter = require('./writer/uint8ArrayWriter');
+
+
+/**
+ * Transform an integer into a string in hexadecimal.
+ * @private
+ * @param {number} dec the number to convert.
+ * @param {number} bytes the number of bytes to generate.
+ * @returns {string} the result.
+ */
+var decToHex = function(dec, bytes) {
+    var hex = "",
+        i;
+    for (i = 0; i < bytes; i++) {
+        hex += String.fromCharCode(dec & 0xff);
+        dec = dec >>> 8;
+    }
+    return hex;
+};
+
+/**
+ * Generate the various parts used in the construction of the final zip file.
+ * @param {string} name the file name.
+ * @param {ZipObject} file the file content.
+ * @param {JSZip.CompressedObject} compressedObject the compressed object.
+ * @param {number} offset the current offset from the start of the zip file.
+ * @return {object} the zip parts.
+ */
+var generateZipParts = function(name, file, compressedObject, offset) {
+    var data = compressedObject.compressedContent,
+        utfEncodedFileName = utils.transformTo("string", utf8.utf8encode(file.name)),
+        comment = file.comment || "",
+        utfEncodedComment = utils.transformTo("string", utf8.utf8encode(comment)),
+        useUTF8ForFileName = utfEncodedFileName.length !== file.name.length,
+        useUTF8ForComment = utfEncodedComment.length !== comment.length,
+        o = file.options,
+        dosTime,
+        dosDate,
+        extraFields = "",
+        unicodePathExtraField = "",
+        unicodeCommentExtraField = "",
+        dir, date;
+
+
+    // handle the deprecated options.dir
+    if (file._initialMetadata.dir !== file.dir) {
+        dir = file.dir;
+    } else {
+        dir = o.dir;
+    }
+
+    // handle the deprecated options.date
+    if(file._initialMetadata.date !== file.date) {
+        date = file.date;
+    } else {
+        date = o.date;
+    }
+
+    // date
+    // @see http://www.delorie.com/djgpp/doc/rbinter/it/52/13.html
+    // @see http://www.delorie.com/djgpp/doc/rbinter/it/65/16.html
+    // @see http://www.delorie.com/djgpp/doc/rbinter/it/66/16.html
+
+    dosTime = date.getHours();
+    dosTime = dosTime << 6;
+    dosTime = dosTime | date.getMinutes();
+    dosTime = dosTime << 5;
+    dosTime = dosTime | date.getSeconds() / 2;
+
+    dosDate = date.getFullYear() - 1980;
+    dosDate = dosDate << 4;
+    dosDate = dosDate | (date.getMonth() + 1);
+    dosDate = dosDate << 5;
+    dosDate = dosDate | date.getDate();
+
+    if (useUTF8ForFileName) {
+        // set the unicode path extra field. unzip needs at least one extra
+        // field to correctly handle unicode path, so using the path is as good
+        // as any other information. This could improve the situation with
+        // other archive managers too.
+        // This field is usually used without the utf8 flag, with a non
+        // unicode path in the header (winrar, winzip). This helps (a bit)
+        // with the messy Windows' default compressed folders feature but
+        // breaks on p7zip which doesn't seek the unicode path extra field.
+        // So for now, UTF-8 everywhere !
+        unicodePathExtraField =
+            // Version
+            decToHex(1, 1) +
+            // NameCRC32
+            decToHex(crc32(utfEncodedFileName), 4) +
+            // UnicodeName
+            utfEncodedFileName;
+
+        extraFields +=
+            // Info-ZIP Unicode Path Extra Field
+            "\x75\x70" +
+            // size
+            decToHex(unicodePathExtraField.length, 2) +
+            // content
+            unicodePathExtraField;
+    }
+
+    if(useUTF8ForComment) {
+
+        unicodeCommentExtraField =
+            // Version
+            decToHex(1, 1) +
+            // CommentCRC32
+            decToHex(this.crc32(utfEncodedComment), 4) +
+            // UnicodeName
+            utfEncodedComment;
+
+        extraFields +=
+            // Info-ZIP Unicode Path Extra Field
+            "\x75\x63" +
+            // size
+            decToHex(unicodeCommentExtraField.length, 2) +
+            // content
+            unicodeCommentExtraField;
+    }
+
+    var header = "";
+
+    // version needed to extract
+    header += "\x0A\x00";
+    // general purpose bit flag
+    // set bit 11 if utf8
+    header += (useUTF8ForFileName || useUTF8ForComment) ? "\x00\x08" : "\x00\x00";
+    // compression method
+    header += compressedObject.compression.magic;
+    // last mod file time
+    header += decToHex(dosTime, 2);
+    // last mod file date
+    header += decToHex(dosDate, 2);
+    // crc-32
+    header += decToHex(compressedObject.crc32, 4);
+    // compressed size
+    header += decToHex(compressedObject.compressedSize, 4);
+    // uncompressed size
+    header += decToHex(compressedObject.uncompressedSize, 4);
+    // file name length
+    header += decToHex(utfEncodedFileName.length, 2);
+    // extra field length
+    header += decToHex(extraFields.length, 2);
+
+
+    var fileRecord = signature.LOCAL_FILE_HEADER + header + utfEncodedFileName + extraFields;
+
+    var dirRecord = signature.CENTRAL_FILE_HEADER +
+    // version made by (00: DOS)
+    "\x14\x00" +
+    // file header (common to file and central directory)
+    header +
+    // file comment length
+    decToHex(utfEncodedComment.length, 2) +
+    // disk number start
+    "\x00\x00" +
+    // internal file attributes TODO
+    "\x00\x00" +
+    // external file attributes
+    (dir === true ? "\x10\x00\x00\x00" : "\x00\x00\x00\x00") +
+    // relative offset of local header
+    decToHex(offset, 4) +
+    // file name
+    utfEncodedFileName +
+    // extra field
+    extraFields +
+    // file comment
+    utfEncodedComment;
+
+    return {
+        fileRecord: fileRecord,
+        dirRecord: dirRecord,
+        compressedObject: compressedObject
+    };
+};
+
+/**
+ * Generate the EOCD record.
+ * @param {Number} entriesCount the number of entries in the zip file.
+ * @param {Number} centralDirLength the length (in bytes) of the central dir.
+ * @param {Number} localDirLength the length (in bytes) of the local dir.
+ * @param {String} utfEncodedComment the zip file comment as a binary string.
+ * @return {String} the EOCD record.
+ */
+var generateCentralDirectoryEnd = function (entriesCount, centralDirLength, localDirLength, utfEncodedComment) {
+    var dirEnd = "";
+
+    // end of central dir signature
+    dirEnd = signature.CENTRAL_DIRECTORY_END +
+        // number of this disk
+        "\x00\x00" +
+        // number of the disk with the start of the central directory
+        "\x00\x00" +
+        // total number of entries in the central directory on this disk
+        decToHex(entriesCount, 2) +
+        // total number of entries in the central directory
+        decToHex(entriesCount, 2) +
+        // size of the central directory   4 bytes
+        decToHex(centralDirLength, 4) +
+        // offset of start of central directory with respect to the starting disk number
+        decToHex(localDirLength, 4) +
+        // .ZIP file comment length
+        decToHex(utfEncodedComment.length, 2) +
+        // .ZIP file comment
+        utfEncodedComment;
+
+    return dirEnd;
+};
+
+var getCompression = function (fileCompression, zipCompression) {
+
+    var compressionName = fileCompression || zipCompression;
+    var compression = compressions[compressionName];
+    if (!compression) {
+        throw new Error(compressionName + " is not a valid compression method !");
+    }
+    return compression;
+};
+/**
+ * Compress every entries in the given object.
+ * @param {Object} files the object containing the files, one file per attribute of the object.
+ * @param {String} zipCompression the upper case name of the compression to use for the zip file.
+ * @return {Array} an array of compressed `ZipObject`s.
+ */
+exports.prepareCompressedEntries = function (files, zipCompression) {
+    var compressedEntries = [];
+    for (var name in files) {
+        if (!files.hasOwnProperty(name)) {
+            continue;
+        }
+        var file = files[name];
+
+        var compression = getCompression(file.options.compression, zipCompression);
+        file._compress(compression);
+        compressedEntries.push(file);
+    }
+    return compressedEntries;
+};
+
+/**
+ * Generate a zip file with the given compressed `ZipObject`s.
+ * @param {String} type the lower case name of the output type.
+ * @param {Array} entries an array of compressed `ZipObject`s.
+ * @param {String} comment the comment of the zip file.
+ * @return {String|Uint8Array|ArrayBuffer|Buffer|Blob} the zip file
+ */
+exports.generateZipFile = function(type, entries, comment) {
+
+    var zipData = [],
+        localDirLength = 0,
+        centralDirLength = 0,
+        writer, i, len,
+        utfEncodedComment = utils.transformTo("string", utf8.utf8encode(comment));
+
+    // first, generate all the zip parts.
+    for(i = 0; i < entries.length; i++) {
+        var file = entries[i];
+
+        var compressedObject = file._data;
+
+        var zipPart = generateZipParts(name, file, compressedObject, localDirLength);
+        localDirLength += zipPart.fileRecord.length + compressedObject.compressedSize;
+        centralDirLength += zipPart.dirRecord.length;
+        zipData.push(zipPart);
+    }
+
+    var dirEnd = generateCentralDirectoryEnd(zipData.length, centralDirLength, localDirLength, utfEncodedComment);
+
+    // we have all the parts (and the total length)
+    // time to create a writer !
+    if(type === "string" || type === "base64") {
+        writer = new StringWriter(localDirLength + centralDirLength + dirEnd.length);
+    }else{
+        writer = new Uint8ArrayWriter(localDirLength + centralDirLength + dirEnd.length);
+    }
+
+    for (i = 0; i < zipData.length; i++) {
+        writer.append(zipData[i].fileRecord);
+        writer.append(zipData[i].compressedObject.compressedContent);
+    }
+    for (i = 0; i < zipData.length; i++) {
+        writer.append(zipData[i].dirRecord);
+    }
+
+    writer.append(dirEnd);
+
+    var zip = writer.finalize();
+
+    switch(type) {
+        case "blob" :
+            return utils.arrayBuffer2Blob(utils.transformTo("arraybuffer", zip));
+        case "base64" :
+            return base64.encode(zip);
+        default :
+            return utils.transformTo(type, zip);
+    }
+
+};
